Adapt APIRequestFactory, APIClient and APITestCase to DTS.
