+++ tenant_schemas/test/drf/__init__.py
@@ -1,0 +1,1 @@
+__author__ = 'roba'
